requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
chardet==5.0.0
distlib==0.3.5
filelock==3.7.1
idna==3.3
parse==1.19.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.11
virtualenv==20.16.3
psycopg2-binary==2.9.3